backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain-anthropic==0.2.4
langsmith==0.1.129
httpx==0.27.2
web3==7.2.0
supabase==2.7.4
pandas==2.2.3
python-dotenv==1.0.1
